api/requirements.txt
flask==2.3.2
joblib==1.3.1
scikit-learn==1.3.0
python-dotenv==1.0.0